SignUp and Login Working
the two main functions in the folder management now are newUser and
updateUser. newUser is to be used only when signing up. updateUser takes
a user class, checks if the new added email doesn't exist (the last
email in the list is assumed to be the update), looks for its id in the
saved users, then updates the user in the index.json .